+++ IMSWebApi/ViewModel/GoodIssueNote/VMTrnGoodIssueNoteItem.cs
@@ -24,7 +24,7 @@ namespace IMSWebApi.ViewModel
         public Nullable<decimal> issuedQuantity { get; set; }
         public decimal rate { get; set; }
         public Nullable<decimal> discountPercentage { get; set; }
-        public int amount { get; set; }
+        public long amount { get; set; }
         public string status { get; set; }
         public Nullable<System.DateTime> statusChangeDate { get; set; }
