+++ IMSWebApi/ViewModel/MaterialQuotation/VMTrnMaterialQuotationItem.cs
@@ -31,9 +31,9 @@ namespace IMSWebApi.ViewModel
         public string orderType { get; set; }
         public decimal rate { get; set; }
         public decimal discountPercentage { get; set; }
-        public int amount { get; set; }
+        public long amount { get; set; }
         public decimal rateWithGST { get; set; }
-        public int amountWithGST { get; set; }
+        public long amountWithGST { get; set; }
         public Nullable<int> gst { get; set; }
         public string status { get; set; }
